backend/sales/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import CustomerViewSet, SalesOrderViewSet

router = DefaultRouter()
router.register(r'customers', CustomerViewSet)
router.register(r'orders', SalesOrderViewSet, basename='salesorder')

urlpatterns = [
    path('', include(router.urls)),
]

